refactor: fetch_join_columns
